commented out failing code, and added some fixes in Pcl

+++ src/main/cpp/modules/Pcl.h
@@ -29,7 +29,7 @@ private:
     const Accessor* sloFlagsAccessor;
 
 	size_t getIndex(long k, long l, long m) const;
-	uint16_t getValue(size_t index, int64_t olcFlags, int16_t slnFlags, int16_t sloFlags) const;
+	uint16_t getValue(size_t index, uint32_t olcFlags, uint8_t slnFlags, uint8_t sloFlags) const;
 	const Accessor& getSourceAccessor(Context& context, string variableName, string sourceSegmentId);
 	void setUpSourceAccessors(Context & context);
     void setUpSegment(Context & context);
